requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
sqlmodel==0.0.14
psycopg2-binary==2.9.9
alembic==1.13.1
python-dotenv==1.0.0
httpx==0.25.2
pydantic-settings==2.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
requests==2.31.0
plexapi==4.15.7